Move common Offense/DefenseAaRoll classes to its own class
