+++ game-core/src/main/java/games/strategy/triplea/delegate/power/calculator/AaDefenseCombatValue.java
@@ -29,7 +29,7 @@ class AaDefenseCombatValue implements CombatValue {
 
   @Override
   public StrengthAndRollCalculator getRoll() {
-    return new AaDefenseRoll(supportFromFriends, supportFromEnemies);
+    return new AaRoll(supportFromFriends, supportFromEnemies);
   }
 
   @Override
@@ -40,26 +40,6 @@ class AaDefenseCombatValue implements CombatValue {
   @Override
   public boolean isDefending() {
     return true;
-  }
-
-  static class AaDefenseRoll extends StrengthAndRollCalculator {
-
-    AaDefenseRoll(final AvailableSupports friendlySupport, final AvailableSupports enemySupport) {
-      super(friendlySupport, enemySupport);
-    }
-
-    @Override
-    public int getValue(final Unit unit) {
-      return RollValue.of(unit.getUnitAttachment().getMaxAaAttacks())
-          .add(addSupport(unit, friendlySupportTracker))
-          .add(addSupport(unit, enemySupportTracker))
-          .getValue();
-    }
-
-    @Override
-    protected Predicate<UnitSupportAttachment> getRuleFilter() {
-      return UnitSupportAttachment::getAaRoll;
-    }
   }
 
   static class AaDefenseStrength extends StrengthAndRollCalculator {
